Add an axds.valid_range_test supporting datetime-like objects
This is a gross range test but accepts the valid range as input rather
than the bounds. The results are the same and this test could be replaced
with similar gross_range_test calls but I was finding it much easier to
use an inclusive acceptable range than compute the non-acceptable range.
This test uses the same standard_name as gross_range_test for metadata
and documentation purposes.